skalp_bot/requirements.txt
numpy
pandas
pyyaml
ccxt
matplotlib
